requirements.txt
streamlit==1.32.0
langchain==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.3
langchain-community==0.3.14
python-dotenv==1.0.1
typing-extensions>=4.7.0 